Alter grid row height for the project cards
Instead of using fractional units for the grids (fr), the project cards
use viewport height (vh) for their height. This is because with fr,
the height of grid item expands when the text content inside of it is
too large. As it expands, the grid item will still be treated as if
it is 1fr, so others project cards will also expand even if the text
isn't overflowing. With vh, the project card has a more defined height
limit, and the overflow property can be used.